Ristrutturazione

+++ list/src/main/java/it/unicam/cs/pa/logo/model/Direction.java
@@ -10,7 +10,7 @@ public interface Direction {
      *
      * @return il valore della direzione
      */
-    Integer getValue();
+    double getValue();
 
     /**
      * Incrementa il valore della direzione
